Added multiple functionalities to the spectrometer UI
Added the connection and disconnection buttons and connected their signals.

Added a status label to let user know if the spectrometer is connected or disconnected.

Added double spin box to let user input the wanted exposure time.

Added a measure button and connected its signal.

Added a widget for plotting the measured spectrum using matplotlib.

A toolbar was added to the plot.